bot/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiohttp-sse-client==0.2.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.5
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
multidict==6.1.0
propcache==0.2.0
pycparser==2.22
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyNaCl==1.5.0
python-dotenv==1.0.1
python-telegram-bot==21.7
pytonconnect==0.3.1
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
yarl==1.18.0
